% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cTU_curation.R
\name{gen_cTU_data}
\alias{gen_cTU_data}
\title{Generate positive and negative TU training datasets}
\usage{
gen_cTU_data(file_RNAseqSignals, fileNamePrefix, file_gff, genomeFile)
}
\arguments{
\item{file_RNAseqSignals}{The file of .NA file}

\item{fileNamePrefix}{The prefix of output file name}

\item{file_gff}{The .gff file of reference genome}

\item{genomeFile}{Reference genome sequence file}
}
\description{
Generate positive and negative TU training datasets
}
