% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NA.R
\name{gen_NA}
\alias{gen_NA}
\title{Generation of .NA file for data training}
\usage{
gen_NA(bamfile, ref_genome_name, output_prefix)
}
\arguments{
\item{bamfile}{The result of alignment in bam format.}

\item{ref_genome_name}{The file name of reference genome such as NC_000913.3}

\item{output_prefix}{The prefix of output file name.}
}
\description{
Generation of .NA file for data training
}
