% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.R
\name{generate_GTF_training}
\alias{generate_GTF_training}
\title{Generate GTF file for training.
It will generate a GTF (mocked.GTF) for training in the working directory.}
\usage{
generate_GTF_training(genomelength)
}
\arguments{
\item{genomelength}{The length of genome}
}
\value{
GTF file for training
}
\description{
Generate GTF file for training.
It will generate a GTF (mocked.GTF) for training in the working directory.
}
