% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{TU_SVM}
\alias{TU_SVM}
\title{Using Support Vector Machine to train and generate TU prediction results}
\usage{
TU_SVM(positive_training, negative_training, positive_strand_testing,
  negative_strand_testing, file_RNAseqSignals, file_gff, output_prefix,
  genome_name)
}
\arguments{
\item{positive_training}{The file of positive training dataset.}

\item{negative_training}{The file of negative training dataset.}

\item{positive_strand_testing}{The file of positive strand testing dataset.}

\item{negative_strand_testing}{The file of negative strand testing dataset.}

\item{file_RNAseqSignals}{The .NA file generated by previous step}

\item{file_gff}{The .gff file of reference genome}

\item{output_prefix}{The prefix of output file name}

\item{genome_name}{The file of referecne genome}
}
\description{
Using Support Vector Machine to train and generate TU prediction results
}
